password encryption + db user save
