+++ pom.xml
@@ -63,6 +63,11 @@
 			<version>2.10.0</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-crypto</artifactId>
+		</dependency>
+
     </dependencies>
 
 	<build>
