+++ src/main/java/com/hotelcalifornia/hotel/repository/UserRepository.java
@@ -12,5 +12,5 @@ public interface UserRepository extends CrudRepository<User, Long> {
     List<User> findAll();
     User findById(long id);
     User findByUserName(String username);
-    User findByEmailAddress(String email);
+//    User findByEmailAddress(String email);
 }
